refactor(capabilities): extract the http client runtime half and evacuate both http test modules
Apply the ADR-0122 identity/runtime split to the `aether.http` egress client
the way `http/server` already does it: convert `HttpCapability` from the
impl-hosted `#[actor(singleton)] impl NativeActor` form to the struct-hosted
`#[actor(singleton)] pub struct HttpCapability;` form, and move the
`aether_substrate`- / `ureq`-typed runtime surface — `HttpCapabilityState`, the
adapter abstraction (`FetchRequest` / `FetchResponse` / `HttpAdapter` /
`DisabledHttpAdapter`), the `UreqHttpAdapter` stack, and the `#[runtime] impl
NativeActor` handler bodies — into a new `client/runtime.rs` behind the one
`feature = "runtime"` gate. `client/mod.rs` keeps only the identity ZST, the
`HttpMailboxExt` facade, the `DEFAULT_*` consts, and the config re-exports, so
the always-on identity markers stay always-on and a transport-only build never
names the state type.

Evacuate the server's inline `test_handlers` + `tests` modules to a sibling
`server/tests.rs`, per the `rpc/server/tests.rs` precedent. The client's tests
drive `on_fetch` directly (a runtime-module-private method), so they live
alongside it in `runtime.rs`, matching the `gemini` / `ui` split caps.

Behavior-preserving: no wire change, no kind change, no ADR.

Closes #2476
